+++ dramatiq/brokers/redis.py
@@ -18,6 +18,7 @@
 import glob
 import random
 import time
+import warnings
 from os import path
 from uuid import uuid4
 
@@ -73,6 +74,8 @@ class RedisBroker(Broker):
         offline.
       dead_message_ttl(int): The amount of time (in ms) that
         dead-lettered messages are kept in Redis for.
+      requeue_deadline(int): Deprecated.  Does nothing.
+      requeue_interval(int): Deprecated.  Does nothing.
       \**parameters(dict): Connection parameters are passed directly
         to :class:`redis.StrictRedis`.
 
@@ -85,12 +88,18 @@ class RedisBroker(Broker):
             maintenance_chance=DEFAULT_MAINTENANCE_CHANCE,
             heartbeat_timeout=DEFAULT_HEARTBEAT_TIMEOUT,
             dead_message_ttl=DEFAULT_DEAD_MESSAGE_TTL,
+            requeue_deadline=None,
+            requeue_interval=None,
             **parameters
     ):
         super().__init__(middleware=middleware)
 
         if url:
             parameters["connection_pool"] = redis.ConnectionPool.from_url(url)
+
+        if requeue_deadline or requeue_interval:
+            message = "requeue_{deadline,interval} have been deprecated and no longer do anything"
+            warnings.warn(message, DeprecationWarning, stacklevel=2)
 
         self.broker_id = str(uuid4())
         self.namespace = namespace